Assignment2/requirements.txt
gudhi==3.7.1
matplotlib==3.7.3
networkx==3.1
numpy==1.23.5
param==1.13.0
persim==0.3.1
rsa==4.9
scipy==1.10.1
yacs==0.1.8
pytorch_geometric==2.5.3